# src/main/resources/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "接口说明文档",
    "version" : "v1",
    "title" : "接口说明文档"
  },
  "host" : "localhost:8080",
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/order/chart/{ordernumber}" : {
      "get" : {
        "summary" : "获取订单图表数据",
        "description" : "",
        "operationId" : "orderChart",
        "parameters" : [ {
          "name" : "ordernumber",
          "in" : "path",
          "description" : "订单编号",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
    "/order/check/{ordernumber}" : {
      "put" : {
        "summary" : "审核单个订单",
        "description" : "",
        "operationId" : "checkOrder",
        "parameters" : [ {
          "name" : "ordernumber",
          "in" : "path",
          "description" : "订单编号",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/HttpServletRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
    "/order/confirm/{ordernumber}" : {
      "put" : {
        "summary" : "确认收货",
        "description" : "",
        "operationId" : "confirmRecOrder",
        "parameters" : [ {
          "name" : "ordernumber",
          "in" : "path",
          "description" : "订单编号",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
    "/order/delete/{ordernumber}" : {
      "delete" : {
        "summary" : "删除对应的订单",
        "description" : "",
        "operationId" : "deleteOrder",
        "parameters" : [ {
          "name" : "ordernumber",
          "in" : "path",
          "description" : "订单编号",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
    "/order/detail/{ordernumber}" : {
      "get" : {
        "summary" : "跳转订单详情界面",
        "description" : "",
        "operationId" : "showOrderDeatil",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Model"
          }
        }, {
          "name" : "ordernumber",
          "in" : "path",
          "description" : "订单编号",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/order/list/{ordernumber}" : {
      "get" : {
        "summary" : "获取订单详细数据",
        "description" : "",
        "operationId" : "orderList",
        "parameters" : [ {
          "name" : "ordernumber",
          "in" : "path",
          "description" : "订单编号",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/RestMessage"
            }
          }
        }
      }
    },
    "/order/reject/{ordernumber}" : {
      "put" : {
        "summary" : "拒绝收货",
        "description" : "",
        "operationId" : "rejectRecOrder",
        "parameters" : [ {
          "name" : "ordernumber",
          "in" : "path",
          "description" : "订单编号",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
...
    "/orders/check" : {
      "put" : {
        "summary" : "批量审核订单",
        "description" : "",
        "operationId" : "checkOrders",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "待确认的订单对象列表",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/OrderFood"
            }
          }
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/HttpServletRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
    "/orders/confirm" : {
      "put" : {
        "summary" : "批量确认收货",
        "description" : "",
        "operationId" : "checkOrder",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "选取的订单对象列表",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/OrderFood"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
...
    "/preorder/make" : {
      "post" : {
        "summary" : "预购请求生成订单",
        "description" : "",
        "operationId" : "makeOrder",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "预购请求列表",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/AskedFood"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "AskedFood" : {
      "type" : "object",
      "properties" : {
        "foodname" : {
          "type" : "string"
        },
        "fooduid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "askedamount" : {
          "type" : "string"
        },
        "finishtime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "ordernumber" : {
          "type" : "integer",
          "format" : "int32"
        },
        "status" : {
          "type" : "string"
        }
      }
    },
    "BufferedReader" : {
      "type" : "object"
    },
    "Cookie" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        },
        "comment" : {
          "type" : "string"
        },
        "domain" : {
          "type" : "string"
        },
        "maxAge" : {
          "type" : "integer",
          "format" : "int32"
        },
        "path" : {
          "type" : "string"
        },
        "secure" : {
          "type" : "boolean",
          "default" : false
        },
        "version" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "Enumeration" : {
      "type" : "object"
    },
    "HttpServletRequest" : {
      "type" : "object",
      "properties" : {
        "method" : {
          "type" : "string"
        },
        "requestedSessionIdValid" : {
          "type" : "boolean",
          "default" : false
        },
        "pathTranslated" : {
          "type" : "string"
        },
        "requestedSessionIdFromCookie" : {
          "type" : "boolean",
          "default" : false
        },
        "requestedSessionIdFromURL" : {
          "type" : "boolean",
          "default" : false
        },
        "requestedSessionIdFromUrl" : {
          "type" : "boolean",
          "default" : false
        },
        "requestedSessionId" : {
          "type" : "string"
        },
        "session" : {
          "$ref" : "#/definitions/HttpSession"
        },
        "userPrincipal" : {
          "$ref" : "#/definitions/Principal"
        },
        "contextPath" : {
          "type" : "string"
        },
        "requestURL" : {
          "$ref" : "#/definitions/StringBuffer"
        },
        "headerNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "remoteUser" : {
          "type" : "string"
        },
        "authType" : {
          "type" : "string"
        },
        "pathInfo" : {
          "type" : "string"
        },
        "servletPath" : {
          "type" : "string"
        },
        "requestURI" : {
          "type" : "string"
        },
        "queryString" : {
          "type" : "string"
        },
        "cookies" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Cookie"
          }
        },
        "scheme" : {
          "type" : "string"
        },
        "inputStream" : {
          "$ref" : "#/definitions/ServletInputStream"
        },
        "protocol" : {
          "type" : "string"
        },
        "reader" : {
          "$ref" : "#/definitions/BufferedReader"
        },
        "contentLength" : {
          "type" : "integer",
          "format" : "int32"
        },
        "contentType" : {
          "type" : "string"
        },
        "locale" : {
          "$ref" : "#/definitions/Locale"
        },
        "attributeNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "characterEncoding" : {
          "type" : "string"
        },
        "parameterMap" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "parameterNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "localName" : {
          "type" : "string"
        },
        "remoteHost" : {
          "type" : "string"
        },
        "remotePort" : {
          "type" : "integer",
          "format" : "int32"
        },
        "serverName" : {
          "type" : "string"
        },
        "localPort" : {
          "type" : "integer",
          "format" : "int32"
        },
        "locales" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "remoteAddr" : {
          "type" : "string"
        },
        "secure" : {
          "type" : "boolean",
          "default" : false
        },
        "localAddr" : {
          "type" : "string"
        },
        "serverPort" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "HttpSession" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "creationTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "attributeNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "servletContext" : {
          "$ref" : "#/definitions/ServletContext"
        },
        "maxInactiveInterval" : {
          "type" : "integer",
          "format" : "int32"
        },
        "sessionContext" : {
          "$ref" : "#/definitions/HttpSessionContext"
        },
        "lastAccessedTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "valueNames" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "new" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
    "HttpSessionContext" : {
      "type" : "object",
      "properties" : {
        "ids" : {
          "$ref" : "#/definitions/Enumeration"
        }
      }
    },
    "Locale" : {
      "type" : "object",
      "properties" : {
        "language" : {
          "type" : "string"
        },
        "country" : {
          "type" : "string"
        },
        "displayCountry" : {
          "type" : "string"
        },
        "displayLanguage" : {
          "type" : "string"
        },
        "displayName" : {
          "type" : "string"
        },
        "displayScript" : {
          "type" : "string"
        },
        "displayVariant" : {
          "type" : "string"
        },
        "extensionKeys" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "iso3Country" : {
          "type" : "string"
        },
        "iso3Language" : {
          "type" : "string"
        },
        "script" : {
          "type" : "string"
        },
        "unicodeLocaleAttributes" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "unicodeLocaleKeys" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "variant" : {
          "type" : "string"
        }
      }
    },
    "Model" : {
      "type" : "object"
    },
    "OrderFood" : {
      "type" : "object",
      "properties" : {
        "ordernumber" : {
          "type" : "integer",
          "format" : "int32"
        },
        "provideraccount" : {
          "type" : "string"
        },
        "totalmoney" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        },
        "ordertime" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
    "Principal" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        }
      }
    },
...
    "RestMessage" : {
      "type" : "object",
      "properties" : {
        "count" : {
          "type" : "string",
          "description" : "数据数量"
        },
        "data" : {
          "type" : "object",
          "description" : "返回数据"
        },
        "code" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "返回编号"
        },
        "msg" : {
          "type" : "string",
          "description" : "返回信息"
        }
      },
      "description" : "返回信息"
    },
    "ServletContext" : {
      "type" : "object",
      "properties" : {
        "attributeNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "initParameterNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "servletContextName" : {
          "type" : "string"
        },
        "majorVersion" : {
          "type" : "integer",
          "format" : "int32"
        },
        "minorVersion" : {
          "type" : "integer",
          "format" : "int32"
        },
        "contextPath" : {
          "type" : "string"
        },
        "servlets" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "servletNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "serverInfo" : {
          "type" : "string"
        }
      }
    },
    "ServletInputStream" : {
      "type" : "object"
    },
    "StringBuffer" : {
      "type" : "object"
    },
...
  }
}